Fix type validation in DevicePropertyValue
Grabs DeviceType of given device, use coerce_tree gem function to get
all parents, and check if it matches given property.

Right now it's taking 3 querys to acomplish this so I'll try to improve
it.

Not tested because coreced_tree is not optimizaed for minitest, will try
to find workaround.